requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=0_gnu
aioeasywebdav=2.4.0=py37_1000
aiohttp=3.6.2=py37h516909a_0
appdirs=1.4.3=py_1
async-timeout=3.0.1=py_1000
attrs=19.3.0=py_0
bcrypt=3.1.7=py37h516909a_0
biopython=1.76=py37h516909a_0
boto3=1.12.32=pyh9f0ad1d_0
botocore=1.15.32=pyh9f0ad1d_0
bzip2=1.0.8=h516909a_2
ca-certificates=2019.11.28=hecc5488_0
cachetools=3.1.1=py_0
cairo=1.16.0=hcf35c78_1003
cd-hit=4.8.1=h8b12597_3
certifi=2019.11.28=py37hc8dfbb8_1
cffi=1.14.0=py37hd463f26_0
chardet=3.0.4=py37hc8dfbb8_1006
configargparse=1.1=pyh8c360ce_0
crc32c=2.0=py37h516909a_0
cryptography=2.8=py37hb09aad4_2
curl=7.68.0=hf8cf82a_0
datrie=0.8.2=py37h8f50634_0
decorator=4.4.2=py_0
docutils=0.15.2=py37_0
dropbox=9.4.0=py_0
entrezpy=2.0.3=py_0
expat=2.2.9=he1b5a44_2
fftw=3.3.8=mpi_mpich_h3f9e1be_1010
filechunkio=1.8=py_2
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.1=he06d7ca_0
ftputil=3.4=py_0
gdk-pixbuf=2.36.12=h3f25603_1005
gettext=0.19.8.1=hc5be6a0_1002
ghostscript=9.22=hf484d3e_1001
giflib=5.1.7=h516909a_1
gitdb=4.0.2=py_0
gitpython=3.1.0=py_0
glib=2.58.3=py37he00f558_1003
gobject-introspection=1.58.2=py37h619baee_1004
google-api-core=1.16.0=py37_1
google-auth=1.12.0=pyh9f0ad1d_0
google-cloud-core=1.3.0=py_0
google-cloud-storage=1.26.0=py_0
google-resumable-media=0.5.0=py_1
googleapis-common-protos=1.51.0=py37_1
graphite2=1.3.13=he1b5a44_1001
graphviz=2.38.0=hf68f40c_1011
harfbuzz=2.4.0=h9f30f68_3
hyphy=2.5.2=h7d38c80_0
icu=64.2=he1b5a44_1
idna=2.9=py_1
imagemagick=7.0.8_54=pl526h39023e4_0
importlib-metadata=1.6.0=py37hc8dfbb8_0
importlib_metadata=1.6.0=0
ipython_genutils=0.2.0=py_1
iqtree=1.6.12=he513fc3_1
jbig=2.1=h14c3975_2001
jinja2=2.11.1=py_0
jmespath=0.9.5=py_0
jpeg=9c=h14c3975_1001
jsonschema=3.2.0=py37hc8dfbb8_1
jupyter_core=4.6.3=py37hc8dfbb8_1
krb5=1.16.4=h2fd8d38_0
ld_impl_linux-64=2.34=h53a641e_0
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libcroco=0.6.13=h8d621e5_0
libcurl=7.68.0=hda55be3_0
libedit=3.1.20170329=hf8c457e_1001
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.2.0=h24d8f2e_2
libgfortran-ng=7.3.0=hdf63c60_5
libgomp=9.2.0=h24d8f2e_2
libiconv=1.15=h516909a_1006
liblapack=3.8.0=14_openblas
libopenblas=0.3.7=h5ec1e0e_6
libpng=1.6.37=hed695b0_1
libprotobuf=3.11.4=h8b12597_0
librsvg=2.44.14=h11c8777_0
libssh2=1.8.2=h22169c7_2
libstdcxx-ng=9.2.0=hdf63c60_2
libtiff=4.1.0=hc3755c2_3
libtool=2.4.6=h14c3975_1002
libuuid=2.32.1=h14c3975_1000
libwebp=1.0.2=hf4e8a37_4
libxcb=1.13=h14c3975_1002
libxml2=2.9.10=hee79883_0
llvm-openmp=8.0.1=hc9558a2_0
lz4-c=1.8.3=he1b5a44_1001
mafft=7.455=h516909a_0
markupsafe=1.1.1=py37h8f50634_1
mpi=1.0=mpich
mpich=3.3.2=hc856adb_0
multidict=4.7.5=py37h516909a_0
nbformat=5.0.4=py_0
ncurses=6.1=hf484d3e_1002
networkx=2.4=py_1
numpy=1.18.1=py37h8960a57_1
openjpeg=2.3.1=h981e76c_3
openmp=8.0.1=0
openssl=1.1.1f=h516909a_0
pal2nal=14.1=pl526_1
pandas=1.0.3=py37h0da4684_0
pango=1.40.14=he7ab937_1005
paramiko=2.7.1=py37_0
pcre=8.44=he1b5a44_0
perl=5.26.2=h516909a_1006
perl-getopt-long=2.50=pl526_1
pip=20.0.2=py_2
pixman=0.38.0=h516909a_1003
pkg-config=0.29.2=h516909a_1006
prank=v.170427=h6bb024c_2
prettytable=0.7.2=py_3
protobuf=3.11.4=py37he1b5a44_0
psutil=5.7.0=py37h8f50634_1
pthread-stubs=0.4=h14c3975_1001
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_0
pygments=2.6.1=py_0
pygraphviz=1.5=py37h8f50634_1002
pynacl=1.3.0=py37h516909a_1001
pyopenssl=19.1.0=py_1
pyrsistent=0.16.0=py37h8f50634_0
pysftp=0.2.9=py_1
pysocks=1.7.1=py37hc8dfbb8_1
python=3.7.6=h8356626_5_cpython
python-dateutil=2.8.1=py_0
python-irodsclient=0.8.2=py_0
python_abi=3.7=1_cp37m
pytz=2019.3=py_0
pyyaml=5.3.1=py37h8f50634_0
ratelimiter=1.2.0=py37_1000
readline=8.0=hf8c457e_0
requests=2.23.0=pyh8c360ce_2
rsa=4.0=py_0
s3transfer=0.3.3=py37hc8dfbb8_0
setuptools=46.1.3=py37hc8dfbb8_0
six=1.14.0=py_1
smmap=3.0.1=py_0
snakemake=5.13.0=0
snakemake-minimal=5.13.0=py_0
sqlite=3.30.1=hcee41ef_0
tk=8.6.10=hed695b0_0
toposort=1.5=py_3
traitlets=4.3.3=py37hc8dfbb8_1
urllib3=1.25.7=py37hc8dfbb8_1
wheel=0.34.2=py_1
wrapt=1.12.1=py37h8f50634_1
xmlrunner=1.7.7=py_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxpm=3.5.13=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-libxt=1.1.5=h516909a_1003
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h516909a_1002
yaml=0.2.2=h516909a_1
yarl=1.3.0=py37h516909a_1000
zipp=3.1.0=py_0
zlib=1.2.11=h516909a_1006
zstd=1.4.4=h3b9ef0a_2
